Struct aws_sdk_codepipeline::client::Client

pub struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

An ergonomic service client for CodePipeline_20150709.

This client allows ergonomic access to a CodePipeline_20150709-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Self

pub fn conf(&self) -> &Config

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn acknowledge_job(&self) -> AcknowledgeJob<C, M, R>

pub fn acknowledge_third_party_job(&self) -> AcknowledgeThirdPartyJob<C, M, R>

pub fn create_custom_action_type(&self) -> CreateCustomActionType<C, M, R>

pub fn create_pipeline(&self) -> CreatePipeline<C, M, R>

pub fn delete_custom_action_type(&self) -> DeleteCustomActionType<C, M, R>

pub fn delete_pipeline(&self) -> DeletePipeline<C, M, R>

pub fn delete_webhook(&self) -> DeleteWebhook<C, M, R>

pub fn deregister_webhook_with_third_party(
&self
) -> DeregisterWebhookWithThirdParty<C, M, R>

pub fn disable_stage_transition(&self) -> DisableStageTransition<C, M, R>

pub fn enable_stage_transition(&self) -> EnableStageTransition<C, M, R>

pub fn get_action_type(&self) -> GetActionType<C, M, R>

pub fn get_job_details(&self) -> GetJobDetails<C, M, R>

pub fn get_pipeline(&self) -> GetPipeline<C, M, R>

pub fn get_pipeline_execution(&self) -> GetPipelineExecution<C, M, R>

pub fn get_pipeline_state(&self) -> GetPipelineState<C, M, R>

pub fn get_third_party_job_details(&self) -> GetThirdPartyJobDetails<C, M, R>

pub fn list_action_executions(&self) -> ListActionExecutions<C, M, R>

pub fn list_action_types(&self) -> ListActionTypes<C, M, R>

pub fn list_pipeline_executions(&self) -> ListPipelineExecutions<C, M, R>

pub fn list_pipelines(&self) -> ListPipelines<C, M, R>

pub fn list_tags_for_resource(&self) -> ListTagsForResource<C, M, R>

pub fn list_webhooks(&self) -> ListWebhooks<C, M, R>

pub fn poll_for_jobs(&self) -> PollForJobs<C, M, R>

pub fn poll_for_third_party_jobs(&self) -> PollForThirdPartyJobs<C, M, R>

pub fn put_action_revision(&self) -> PutActionRevision<C, M, R>

pub fn put_approval_result(&self) -> PutApprovalResult<C, M, R>

pub fn put_job_failure_result(&self) -> PutJobFailureResult<C, M, R>

pub fn put_job_success_result(&self) -> PutJobSuccessResult<C, M, R>

pub fn put_third_party_job_failure_result(
&self
) -> PutThirdPartyJobFailureResult<C, M, R>

pub fn put_third_party_job_success_result(
&self
) -> PutThirdPartyJobSuccessResult<C, M, R>

pub fn put_webhook(&self) -> PutWebhook<C, M, R>

pub fn register_webhook_with_third_party(
&self
) -> RegisterWebhookWithThirdParty<C, M, R>

pub fn retry_stage_execution(&self) -> RetryStageExecution<C, M, R>

pub fn start_pipeline_execution(&self) -> StartPipelineExecution<C, M, R>

pub fn stop_pipeline_execution(&self) -> StopPipelineExecution<C, M, R>

pub fn tag_resource(&self) -> TagResource<C, M, R>

pub fn untag_resource(&self) -> UntagResource<C, M, R>

pub fn update_action_type(&self) -> UpdateActionType<C, M, R>

pub fn update_pipeline(&self) -> UpdatePipeline<C, M, R>

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(conf: Config, conn: C) -> Self

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Self

pub fn from_conf(conf: Config) -> Self