Module aws_sdk_codepipeline::error

Modules

action_not_found_exception

See ActionNotFoundException

action_type_not_found_exception

See ActionTypeNotFoundException

approval_already_completed_exception

See ApprovalAlreadyCompletedException

concurrent_modification_exception

See ConcurrentModificationException

conflict_exception

See ConflictException

duplicated_stop_request_exception

See DuplicatedStopRequestException

invalid_action_declaration_exception

See InvalidActionDeclarationException

invalid_approval_token_exception

See InvalidApprovalTokenException

invalid_arn_exception

See InvalidArnException

invalid_blocker_declaration_exception

See InvalidBlockerDeclarationException

invalid_client_token_exception

See InvalidClientTokenException

invalid_job_exception

See InvalidJobException

invalid_job_state_exception

See InvalidJobStateException

invalid_next_token_exception

See InvalidNextTokenException

invalid_nonce_exception

See InvalidNonceException

invalid_stage_declaration_exception

See InvalidStageDeclarationException

invalid_structure_exception

See InvalidStructureException

invalid_tags_exception

See InvalidTagsException

invalid_webhook_authentication_parameters_exception

See InvalidWebhookAuthenticationParametersException

invalid_webhook_filter_pattern_exception

See InvalidWebhookFilterPatternException

job_not_found_exception

See JobNotFoundException

limit_exceeded_exception

See LimitExceededException

not_latest_pipeline_execution_exception

See NotLatestPipelineExecutionException

output_variables_size_exceeded_exception

See OutputVariablesSizeExceededException

pipeline_execution_not_found_exception

See PipelineExecutionNotFoundException

pipeline_execution_not_stoppable_exception

See PipelineExecutionNotStoppableException

pipeline_name_in_use_exception

See PipelineNameInUseException

pipeline_not_found_exception

See PipelineNotFoundException

pipeline_version_not_found_exception

See PipelineVersionNotFoundException

request_failed_exception

See RequestFailedException

resource_not_found_exception

See ResourceNotFoundException

stage_not_found_exception

See StageNotFoundException

stage_not_retryable_exception

See StageNotRetryableException

too_many_tags_exception

See TooManyTagsException

validation_exception

See ValidationException

webhook_not_found_exception

See WebhookNotFoundException

Structs

AcknowledgeJobError

AcknowledgeThirdPartyJobError

ActionNotFoundException

The specified action cannot be found.

ActionTypeNotFoundException

The specified action type cannot be found.

ApprovalAlreadyCompletedException

The approval action has already been approved or rejected.

ConcurrentModificationException

Unable to modify the tag due to a simultaneous update request.

ConflictException

Your request cannot be handled because the pipeline is busy handling ongoing

CreateCustomActionTypeError

CreatePipelineError

DeleteCustomActionTypeError

DeletePipelineError

DeleteWebhookError

DeregisterWebhookWithThirdPartyError

DisableStageTransitionError

DuplicatedStopRequestException

The pipeline execution is already in a Stopping state. If you already

EnableStageTransitionError

GetActionTypeError

GetJobDetailsError

GetPipelineError

GetPipelineExecutionError

GetPipelineStateError

GetThirdPartyJobDetailsError

InvalidActionDeclarationException

The action declaration was specified in an invalid format.

InvalidApprovalTokenException

The approval request already received a response or has expired.

InvalidArnException

The specified resource ARN is invalid.

InvalidBlockerDeclarationException

Reserved for future use.

InvalidClientTokenException

The client token was specified in an invalid format

InvalidJobException

The job was specified in an invalid format or cannot be found.

InvalidJobStateException

The job state was specified in an invalid format.

InvalidNextTokenException

The next token was specified in an invalid format. Make sure that the next token

InvalidNonceException

The nonce was specified in an invalid format.

InvalidStageDeclarationException

The stage declaration was specified in an invalid format.

InvalidStructureException

The structure was specified in an invalid format.

InvalidTagsException

The specified resource tags are invalid.

InvalidWebhookAuthenticationParametersException

The specified authentication type is in an invalid format.

InvalidWebhookFilterPatternException

The specified event filter rule is in an invalid format.

JobNotFoundException

The job was specified in an invalid format or cannot be found.

LimitExceededException

The number of pipelines associated with the AWS account has exceeded the limit

ListActionExecutionsError

ListActionTypesError

ListPipelineExecutionsError

ListPipelinesError

ListTagsForResourceError

ListWebhooksError

NotLatestPipelineExecutionException

The stage has failed in a later run of the pipeline and the pipelineExecutionId

OutputVariablesSizeExceededException

Exceeded the total size limit for all variables in the pipeline.

PipelineExecutionNotFoundException

The pipeline execution was specified in an invalid format or cannot be found, or an

PipelineExecutionNotStoppableException

Unable to stop the pipeline execution. The execution might already be in a

PipelineNameInUseException

The specified pipeline name is already in use.

PipelineNotFoundException

The pipeline was specified in an invalid format or cannot be found.

PipelineVersionNotFoundException

The pipeline version was specified in an invalid format or cannot be

PollForJobsError

PollForThirdPartyJobsError

PutActionRevisionError

PutApprovalResultError

PutJobFailureResultError

PutJobSuccessResultError

PutThirdPartyJobFailureResultError

PutThirdPartyJobSuccessResultError

PutWebhookError

RegisterWebhookWithThirdPartyError

RequestFailedException

The request failed because of an unknown error, exception, or failure.

ResourceNotFoundException

The resource was specified in an invalid format.

RetryStageExecutionError

StageNotFoundException

The stage was specified in an invalid format or cannot be found.

StageNotRetryableException

Unable to retry. The pipeline structure or stage state might have changed while

StartPipelineExecutionError

StopPipelineExecutionError

TagResourceError

TooManyTagsException

The tags limit for a resource has been exceeded.

UntagResourceError

UpdateActionTypeError

UpdatePipelineError

ValidationException

The validation was specified in an invalid format.

WebhookNotFoundException

The specified webhook was entered in an invalid format or cannot be

Enums

AcknowledgeJobErrorKind

AcknowledgeThirdPartyJobErrorKind

CreateCustomActionTypeErrorKind

CreatePipelineErrorKind

DeleteCustomActionTypeErrorKind

DeletePipelineErrorKind

DeleteWebhookErrorKind

DeregisterWebhookWithThirdPartyErrorKind

DisableStageTransitionErrorKind

EnableStageTransitionErrorKind

GetActionTypeErrorKind

GetJobDetailsErrorKind

GetPipelineErrorKind

GetPipelineExecutionErrorKind

GetPipelineStateErrorKind

GetThirdPartyJobDetailsErrorKind

ListActionExecutionsErrorKind

ListActionTypesErrorKind

ListPipelineExecutionsErrorKind

ListPipelinesErrorKind

ListTagsForResourceErrorKind

ListWebhooksErrorKind

PollForJobsErrorKind

PollForThirdPartyJobsErrorKind

PutActionRevisionErrorKind

PutApprovalResultErrorKind

PutJobFailureResultErrorKind

PutJobSuccessResultErrorKind

PutThirdPartyJobFailureResultErrorKind

PutThirdPartyJobSuccessResultErrorKind

PutWebhookErrorKind

RegisterWebhookWithThirdPartyErrorKind

RetryStageExecutionErrorKind

StartPipelineExecutionErrorKind

StopPipelineExecutionErrorKind

TagResourceErrorKind

UntagResourceErrorKind

UpdateActionTypeErrorKind

UpdatePipelineErrorKind