Module aws_sdk_codepipeline::model

Modules

action_configuration

See ActionConfiguration

action_configuration_property

See ActionConfigurationProperty

action_context

See ActionContext

action_declaration

See ActionDeclaration

action_execution

See ActionExecution

action_execution_detail

See ActionExecutionDetail

action_execution_filter

See ActionExecutionFilter

action_execution_input

See ActionExecutionInput

action_execution_output

See ActionExecutionOutput

action_execution_result

See ActionExecutionResult

action_revision

See ActionRevision

action_state

See ActionState

action_type

See ActionType

action_type_artifact_details

See ActionTypeArtifactDetails

action_type_declaration

See ActionTypeDeclaration

action_type_executor

See ActionTypeExecutor

action_type_id

See ActionTypeId

action_type_identifier

See ActionTypeIdentifier

action_type_permissions

See ActionTypePermissions

action_type_property

See ActionTypeProperty

action_type_settings

See ActionTypeSettings

action_type_urls

See ActionTypeUrls

approval_result

See ApprovalResult

artifact

See Artifact

artifact_detail

See ArtifactDetail

artifact_details

See ArtifactDetails

artifact_location

See ArtifactLocation

artifact_revision

See ArtifactRevision

artifact_store

See ArtifactStore

aws_session_credentials

See AwsSessionCredentials

blocker_declaration

See BlockerDeclaration

current_revision

See CurrentRevision

encryption_key

See EncryptionKey

error_details

See ErrorDetails

execution_details

See ExecutionDetails

execution_trigger

See ExecutionTrigger

executor_configuration

See ExecutorConfiguration

failure_details

See FailureDetails

input_artifact

See InputArtifact

job

See Job

job_data

See JobData

job_details

See JobDetails

job_worker_executor_configuration

See JobWorkerExecutorConfiguration

lambda_executor_configuration

See LambdaExecutorConfiguration

list_webhook_item

See ListWebhookItem

output_artifact

See OutputArtifact

pipeline_context

See PipelineContext

pipeline_declaration

See PipelineDeclaration

pipeline_execution

See PipelineExecution

pipeline_execution_summary

See PipelineExecutionSummary

pipeline_metadata

See PipelineMetadata

pipeline_summary

See PipelineSummary

s3_artifact_location

See S3ArtifactLocation

s3_location

See S3Location

source_revision

See SourceRevision

stage_context

See StageContext

stage_declaration

See StageDeclaration

stage_execution

See StageExecution

stage_state

See StageState

stop_execution_trigger

See StopExecutionTrigger

tag

See Tag

third_party_job

See ThirdPartyJob

third_party_job_data

See ThirdPartyJobData

third_party_job_details

See ThirdPartyJobDetails

transition_state

See TransitionState

webhook_auth_configuration

See WebhookAuthConfiguration

webhook_definition

See WebhookDefinition

webhook_filter_rule

See WebhookFilterRule

Structs

ActionConfiguration

Represents information about an action configuration.

ActionConfigurationProperty

Represents information about an action configuration property.

ActionContext

Represents the context of an action in the stage of a pipeline to a job

ActionDeclaration

Represents information about an action declaration.

ActionExecution

Represents information about the run of an action.

ActionExecutionDetail

Returns information about an execution of an action, including the action execution

ActionExecutionFilter

Filter values for the action execution.

ActionExecutionInput

Input information used for an action execution.

ActionExecutionOutput

Output details listed for an action execution, such as the action execution

ActionExecutionResult

Execution result information, such as the external execution ID.

ActionRevision

Represents information about the version (or revision) of an action.

ActionState

Represents information about the state of an action.

ActionType

Returns information about the details of an action type.

ActionTypeArtifactDetails

Information about parameters for artifacts associated with the action type, such as

ActionTypeDeclaration

The parameters for the action type definition that are provided when the action type

ActionTypeExecutor

The action engine, or executor, for an action type created for a provider, where the

ActionTypeId

Represents information about an action type.

ActionTypeIdentifier

Specifies the category, owner, provider, and version of the action type.

ActionTypePermissions

Details identifying the users with permissions to use the action type.

ActionTypeProperty

Represents information about each property specified in the action configuration, such

ActionTypeSettings

Returns information about the settings for an action type.

ActionTypeUrls

Returns information about URLs for web pages that display to customers as links on the

ApprovalResult

Represents information about the result of an approval request.

Artifact

Represents information about an artifact that is worked on by actions in the

ArtifactDetail

Artifact details for the action execution, such as the artifact location.

ArtifactDetails

Returns information about the details of an artifact.

ArtifactLocation

Represents information about the location of an artifact.

ArtifactRevision

Represents revision details of an artifact.

ArtifactStore

The S3 bucket where artifacts for the pipeline are stored.

AwsSessionCredentials

Represents an AWS session credentials object. These credentials are temporary

BlockerDeclaration

Reserved for future use.

CurrentRevision

Represents information about a current revision.

EncryptionKey

Represents information about the key used to encrypt data in the artifact store,

ErrorDetails

Represents information about an error in AWS CodePipeline.

ExecutionDetails

The details of the actions taken and results produced on an artifact as it passes

ExecutionTrigger

The interaction or event that started a pipeline execution.

ExecutorConfiguration

The action engine, or executor, related to the supported integration model used to

FailureDetails

Represents information about failure details.

InputArtifact

Represents information about an artifact to be worked on, such as a test or build

Job

Represents information about a job.

JobData

Represents other information about a job required for a job worker to complete the

JobDetails

Represents information about the details of a job.

JobWorkerExecutorConfiguration

Details about the polling configuration for the JobWorker action engine,

LambdaExecutorConfiguration

Details about the configuration for the Lambda action engine, or

ListWebhookItem

The detail returned for each webhook after listing webhooks, such as the webhook

OutputArtifact

Represents information about the output of an action.

PipelineContext

Represents information about a pipeline to a job worker.

PipelineDeclaration

Represents the structure of actions and stages to be performed in the

PipelineExecution

Represents information about an execution of a pipeline.

PipelineExecutionSummary

Summary information about a pipeline execution.

PipelineMetadata

Information about a pipeline.

PipelineSummary

Returns a summary of a pipeline.

S3ArtifactLocation

The location of the S3 bucket that contains a revision.

S3Location

The Amazon S3 artifact location for an action's artifacts.

SourceRevision

Information about the version (or revision) of a source artifact that initiated a

StageContext

Represents information about a stage to a job worker.

StageDeclaration

Represents information about a stage and its definition.

StageExecution

Represents information about the run of a stage.

StageState

Represents information about the state of the stage.

StopExecutionTrigger

The interaction that stopped a pipeline execution.

Tag

A tag is a key-value pair that is used to manage the resource.

ThirdPartyJob

A response to a PollForThirdPartyJobs request returned by AWS

ThirdPartyJobData

Represents information about the job data for a partner action.

ThirdPartyJobDetails

The details of a job sent in response to a GetThirdPartyJobDetails

TransitionState

Represents information about the state of transitions between one stage and another

WebhookAuthConfiguration

The authentication applied to incoming webhook trigger requests.

WebhookDefinition

Represents information about a webhook and its definition.

WebhookFilterRule

The event criteria that specify when a webhook notification is sent to your

Enums

ActionCategory

ActionConfigurationPropertyType

ActionExecutionStatus

ActionOwner

ApprovalStatus

ArtifactLocationType

ArtifactStoreType

BlockerType

EncryptionKeyType

ExecutorType

FailureType

JobStatus

PipelineExecutionStatus

StageExecutionStatus

StageRetryMode

StageTransitionType

TriggerType

WebhookAuthenticationType