Module aws_sdk_codepipeline::operation

Structs

AcknowledgeJob

Returns information about a specified job and whether that job has been received by

AcknowledgeThirdPartyJob

Confirms a job worker has received the specified job. Used for partner actions

CreateCustomActionType

Creates a new custom action that can be used in all pipelines associated with the

CreatePipeline

Creates a pipeline.

DeleteCustomActionType

Marks a custom action as deleted. PollForJobs for the custom action

DeletePipeline

Deletes the specified pipeline.

DeleteWebhook

Deletes a previously created webhook by name. Deleting the webhook stops AWS

DeregisterWebhookWithThirdParty

Removes the connection between the webhook that was created by CodePipeline and the

DisableStageTransition

Prevents artifacts in a pipeline from transitioning to the next stage in the

EnableStageTransition

Enables artifacts in a pipeline to transition to a stage in a pipeline.

GetActionType

Returns information about an action type created for an external provider, where the

GetJobDetails

Returns information about a job. Used for custom actions only.

GetPipeline

Returns the metadata, structure, stages, and actions of a pipeline. Can be used to

GetPipelineExecution

Returns information about an execution of a pipeline, including details about

GetPipelineState

Returns information about the state of a pipeline, including the stages and

GetThirdPartyJobDetails

Requests the details of a job for a third party action. Used for partner actions

ListActionExecutions

Lists the action executions that have occurred in a pipeline.

ListActionTypes

Gets a summary of all AWS CodePipeline action types associated with your

ListPipelineExecutions

Gets a summary of the most recent executions for a pipeline.

ListPipelines

Gets a summary of all of the pipelines associated with your account.

ListTagsForResource

Gets the set of key-value pairs (metadata) that are used to manage the

ListWebhooks

Gets a listing of all the webhooks in this AWS Region for this account. The output

PollForJobs

Returns information about any jobs for AWS CodePipeline to act on.

PollForThirdPartyJobs

Determines whether there are any third party jobs for a job worker to act on. Used

PutActionRevision

Provides information to AWS CodePipeline about new revisions to a source.

PutApprovalResult

Provides the response to a manual approval request to AWS CodePipeline. Valid

PutJobFailureResult

Represents the failure of a job as returned to the pipeline by a job worker. Used

PutJobSuccessResult

Represents the success of a job as returned to the pipeline by a job worker. Used

PutThirdPartyJobFailureResult

Represents the failure of a third party job as returned to the pipeline by a job

PutThirdPartyJobSuccessResult

Represents the success of a third party job as returned to the pipeline by a job

PutWebhook

Defines a webhook and returns a unique webhook URL generated by CodePipeline. This

RegisterWebhookWithThirdParty

Configures a connection between the webhook that was created and the external tool

RetryStageExecution

Resumes the pipeline execution by retrying the last failed actions in a stage. You

StartPipelineExecution

Starts the specified pipeline. Specifically, it begins processing the latest commit

StopPipelineExecution

Stops the specified pipeline execution. You choose to either stop the pipeline

TagResource

Adds to or modifies the tags of the given resource. Tags are metadata that can be used

UntagResource

Removes tags from an AWS resource.

UpdateActionType

Updates an action type that was created with any supported integration model, where

UpdatePipeline

Updates a specified pipeline with edits or changes to its structure. Use a JSON